refact(tests): removed code duplication.
